+++ examples/pyinstrument_.py
@@ -37,7 +37,7 @@ if __name__ == "__main__":
     profiler = PyInstrumentProfiler()
     instrumentor = SklearnInstrumentor(instrument=profiler)
 
-    instrumentor.instrument_estimator(
+    instrumentor.instrument_instance(
         classification_model,
         instrument_kwargs={
             "profiler_kwargs": {"interval": 0.001},