+++ sklearn_instrumentation/types.py
@@ -1,0 +1,6 @@
+from typing import Type
+from typing import Union
+
+from sklearn.base import BaseEstimator
+
+Estimator = Union[Type[BaseEstimator], BaseEstimator]